+++ projects/ngx-xchange-ui/src/components/modal-sublayer/modal-sublayer/modal-sublayer.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectorRef, Component, EventEmitter, inject, Inject, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Observable, timeInterval } from 'rxjs';
 type SubLayerType = 'hint' | 'with buttons' | 'success'
 type SubLayerInformation = {
   title?: string;
@@ -13,20 +14,28 @@ type SubLayerInformation = {
   styleUrl: './modal-sublayer.component.css'
 })
 export class ModalSubLayerComponent implements OnInit {
-  @Input() subLayer:SubLayerType = 'hint';
-  @Input() subLayerInformation: SubLayerInformation ={
+  @Input() subLayerObserver:Observable<SubLayerType> = new Observable<SubLayerType>();
+  @Input() subLayerInformationObserver: Observable<SubLayerInformation> = new Observable<SubLayerInformation>()
+  @Input() hintInformationObserver: Observable<string> = new Observable<string>()
+  @Input() isVisible: Observable<boolean> = new Observable<boolean>();
+  @Output() isVisibleChange = new EventEmitter<boolean>();
+  @Input() userDecision: Observable <boolean> = new Observable<boolean>();
+  @Output() userDecisionChange = new EventEmitter<boolean>();
+  public subLayer: SubLayerType = 'hint'
+  public subLayerInformation: SubLayerInformation = {
     title: 'Copy',
     description: 'Do you want to copy this Objective as Identify ecosystem participants - (1)?',
     buttonDescription: 'Copy objective'
   }
-  @Input() hintInformation: string = 'Link copied to clipboard'
-  @Input() isVisible: boolean = false;
-  @Output() isVisibleChange = new EventEmitter<boolean>();
-  @Input() userDecision: boolean = false;
-  @Output() userDecisionChange = new EventEmitter<boolean>();
+  public hintInformation: string = 'Link copied to clipboard';
+  constructor() {
+  }
   
   
   ngOnInit(): void {
+    this.subLayerObserver.subscribe(res => this.subLayer = res)
+    this.subLayerInformationObserver.subscribe(res => this.subLayerInformation = res);
+    this.hintInformationObserver.subscribe(res => this.hintInformation = res)
     if(['hint', 'success'].includes(this.subLayer)){
       setTimeout(() => {
         this.isVisibleChange.emit(false)
@@ -50,5 +59,13 @@ export class ModalSubLayerComponent implements OnInit {
   closeSubLayerUserDecisionAccept(){
     this.userDecisionChange.emit(true)
     this.isVisibleChange.emit(false)
+  }
+  getButtonStyle(){
+    let style = 'filled'
+    const formatString = this.subLayerInformation?.buttonDescription?.toLowerCase()
+    if(formatString?.includes('delete')){
+      style ='cancel_filled'
+    }
+    return style
   }
 }